+++ src/Addons/Fast.php
@@ -1,0 +1,17 @@
+<?php
+
+
+namespace Askoldex\Teletant\Addons;
+
+
+use Askoldex\Teletant\Context;
+
+class Fast
+{
+    public static function answer($text, $keyboard = null)
+    {
+        return function (Context $ctx) use ($text, $keyboard) {
+            $ctx->reply($text, $keyboard);
+        };
+    }
+}