+++ src/Exception/ContextContainerException.php
@@ -1,0 +1,12 @@
+<?php
+
+
+namespace Askoldex\Teletant\Exception;
+
+
+use Askoldex\Teletant\Interfaces\ContextContainerExceptionInterface;
+
+class ContextContainerException extends TeletantException implements ContextContainerExceptionInterface
+{
+
+}