+++ src/Interfaces/ContextContainerExceptionInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace Askoldex\Teletant\Interfaces;
+
+
+interface ContextContainerExceptionInterface extends TeletantExceptionInterface
+{
+
+}