+++ src/Interfaces/TeletantExceptionInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace Askoldex\Teletant\Interfaces;
+
+
+interface TeletantExceptionInterface
+{
+
+}