+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.3.0] - 2020-04-06
+### Added
+
+- Context Container (DI + Service Container)
+- Default property "resize_keyboard" for Menux
+- Multiple fields ("|" separated) for onMessage, onUpdate events.
+
+### Fixed
+
+- Menux property bug
+
 
 ## [1.2.0] - 2020-03-30
 ### Added